+++ pgdrive/tests/test_functionality/test_random_engine.py
@@ -102,6 +102,49 @@ def test_random_traffic():
         env.close()
 
 
+def test_random_lane_width():
+    env = PGDriveEnv({
+        "environment_num": 5,
+        "traffic_density": .2,
+        "traffic_mode": "trigger",
+        "start_seed": 12,
+        "random_lane_width": True,
+        "load_map_from_json": False
+    })
+    o = env.reset(force_seed=12)
+    old_config_1 = env.vehicle.lane.width
+    env.reset(force_seed=15)
+    old_config_2 = env.vehicle.lane.width
+    env.reset(force_seed=13)
+    env.reset(force_seed=12)
+    new_config = env.vehicle.lane.width
+    assert old_config_1 == new_config
+    env.reset(force_seed=15)
+    new_config = env.vehicle.lane.width
+    assert old_config_2 == new_config
+
+
+def test_random_vehicle_parameter():
+    env = PGDriveEnv({
+        "environment_num": 5,
+        "traffic_density": .2,
+        "traffic_mode": "trigger",
+        "start_seed": 12,
+        "random_agent_model": True
+    })
+    o = env.reset(force_seed=12)
+    old_config_1 = env.vehicle.get_config(True)
+    env.reset(force_seed=15)
+    old_config_2 = env.vehicle.get_config(True)
+    env.reset(force_seed=13)
+    env.reset(force_seed=12)
+    new_config = env.vehicle.get_config(True)
+    assert recursive_equal(old_config_1, new_config)
+    env.reset(force_seed=15)
+    new_config = env.vehicle.get_config(True)
+    assert recursive_equal(old_config_2, new_config)
+
+
 if __name__ == '__main__':
     # test_map_random_seeding()
     test_seeding()
